requirements.txt
absl-py==1.3.0
alabaster==0.7.12
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
astunparse==1.6.3
attrs==22.1.0
audioread==3.0.0
awscli==1.27.20
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
botocore==1.29.20
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.4
configparser==5.3.0
coverage==6.5.0
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docutils==0.16
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flake8==6.0.0
future==0.18.2
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.29
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.51.1
h5py==2.10.0
hyperas==0.4.1
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.1.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
librosa==0.7.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
ordereddict==1.1
packaging==21.3
pandas==1.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.33
protobuf==3.19.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.5.1
pytz==2022.6
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.4.1
Send2Trash==1.8.0
sentry-sdk==1.11.1
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soundfile==0.11.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
subprocess32==3.5.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==2.1.1
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
urllib3==1.26.13
wandb==0.8.36
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.11.0
